Move annotation tests elsewhere
